requirements.txt
python-telegram-bot==20.0
python-dotenv==1.2.2
google-api-python-client==2.195.0
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
requests==2.32.3
